requirements.txt
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
get==2019.4.13
idna==2.9
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.1
opencv-python==4.2.0.32
Pillow==7.0.0
post==2019.4.13
public==2019.4.13
pydicom==1.4.1
pynetdicom==1.4.1
pyparsing==2.4.6
pypng==0.0.20
python-dateutil==2.8.1
query-string==2019.4.13
request==2019.4.13
requests==2.23.0
SimpleITK==1.2.4
six==1.14.0
urllib3==1.25.8
uuid==1.30
validators==0.14.2
wget==3.2
